fix: move _experiments to collection
