+++ rslib/src/backend/config.rs
@@ -57,14 +57,8 @@ impl From<StringKeyProto> for StringKey {
 
 impl crate::services::ConfigService for Collection {
     fn get_config_json(&mut self, input: generic::String) -> Result<generic::Json> {
-        let key = input.val.as_str();
-        let val: Option<Value> = self.get_config_optional(key);
-        let default = match key {
-            "experimentalFeatures" => Some(serde_json::from_str("{}").unwrap()),
-            _ => None,
-        };
-        val.or(default)
-            .or_not_found(key)
+        let val: Option<Value> = self.get_config_optional(input.val.as_str());
+        val.or_not_found(input.val)
             .and_then(|v| serde_json::to_vec(&v).map_err(Into::into))
             .map(Into::into)
     }
